added edit for more fields
