+++ frontend/app/(no-footer)/organiser/event/[id]/page.tsx
@@ -3,13 +3,13 @@
 import ShareModal from "@/components/events/ShareModal";
 import EventDrilldownBanner from "@/components/organiser/EventDrilldownBanner";
 import EventDrilldownCommunicationPage from "@/components/organiser/EventDrilldownCommunicationPage";
-import EventDrilldownDetailsPage from "@/components/organiser/EventDrilldownDetailsPage";
 import EventDrilldownManageAttendeesPage from "@/components/organiser/EventDrilldownManageAttendeesPage";
 import EventDrilldownSettingsPage from "@/components/organiser/EventDrilldownSettingsPage";
 import EventDrilldownSharePage from "@/components/organiser/EventDrilldownSharePage";
 import EventDrilldownSidePanel from "@/components/organiser/EventDrilldownSidePanel";
 import EventDrilldownStatBanner from "@/components/organiser/EventDrilldownStatBanner";
 import OrganiserNavbar from "@/components/organiser/OrganiserNavbar";
+import EventDrilldownDetailsPage from "@/components/organiser/event/details/EventDrilldownDetailsPage";
 import { MobileEventDrilldownNavTabs } from "@/components/organiser/mobile/MobileEventDrilldownNavTabs";
 import { EmptyEventMetadata, EventData, EventId, EventMetadata } from "@/interfaces/EventTypes";
 import { EmptyUserData, UserData } from "@/interfaces/UserTypes";
@@ -38,12 +38,15 @@ export default function EventPage({ params }: EventPageProps) {
   const [eventVacancy, setEventVacancy] = useState<number>(0);
   const [eventDescription, setEventDescription] = useState<string>("");
   const [eventLocation, setEventLocation] = useState<string>("");
+  const [eventSport, setEventSport] = useState<string>("");
   const [eventPrice, setEventPrice] = useState<number>(0);
   const [eventImage, setEventImage] = useState<string>("");
   const [eventAccessCount, setEventAccessCount] = useState<number>(0);
   const [eventCapacity, setEventCapacity] = useState<number>(0);
   const [eventMetadata, setEventMetadata] = useState<EventMetadata>(EmptyEventMetadata);
   const [eventPaused, setEventPaused] = useState<boolean>(false);
+  const [eventRegistrationDeadline, setEventRegistrationDeadline] = useState<Timestamp>(Timestamp.now());
+  const [eventIsActive, setEventIsActive] = useState<boolean>(false);
 
   const router = useRouter();
 
@@ -59,11 +62,14 @@ export default function EventPage({ params }: EventPageProps) {
         setEventVacancy(event.vacancy);
         setEventDescription(event.description);
         setEventLocation(event.location);
+        setEventSport(event.sport);
         setEventPrice(event.price);
         setEventImage(event.image);
         setEventAccessCount(event.accessCount);
         setEventCapacity(event.capacity);
         setEventPaused(event.paused);
+        setEventRegistrationDeadline(event.registrationDeadline);
+        setEventIsActive(event.isActive);
       })
       .finally(async () => {
         await sleep(500);
@@ -138,9 +144,13 @@ export default function EventPage({ params }: EventPageProps) {
                   eventEndDate={eventEndDate}
                   eventDescription={eventDescription}
                   eventLocation={eventLocation}
+                  eventSport={eventSport}
+                  eventCapacity={eventCapacity}
                   eventPrice={eventPrice}
                   eventImage={eventImage}
                   eventId={eventId}
+                  eventRegistrationDeadline={eventRegistrationDeadline}
+                  isActive={eventIsActive}
                   updateData={updateEventById}
                 />
                 <ShareModal eventId={eventId} />
