+++ frontend/app/globals.css
@@ -4,14 +4,14 @@
 
 @layer components {
   .screen-width-primary {
-    @apply w-[400px] sm:w-[500px] md:w-[700px] lg:w-[1000px] xl:w-[1000px] 2xl:w-[1400px];
+    @apply w-[400px] sm:w-[500px] md:w-[700px] lg:w-[1000px] xl:w-[1000px] 2xl:w-[1200px];
   }
   .screen-width-dashboard {
-    @apply w-screen px-3 sm:mx-0 sm:w-[500px] md:w-[700px] lg:w-[1000px] xl:w-[1200px] 2xl:w-[1400px];
+    @apply w-screen px-3 sm:mx-0 sm:w-[500px] md:w-[700px] lg:w-[1000px] xl:w-[1000px] 2xl:w-[1200px];
   }
 }
 @layer base {
-  .list-style-circle{
+  .list-style-circle {
     list-style: circle;
   }
 }
@@ -20,13 +20,13 @@
   @variants responsive {
     /* Hide scrollbar for Chrome, Safari and Opera */
     .no-scrollbar::-webkit-scrollbar {
-        display: none;
+      display: none;
     }
 
     /* Hide scrollbar for IE, Edge and Firefox */
     .no-scrollbar {
-        -ms-overflow-style: none;  /* IE and Edge */
-        scrollbar-width: none;  /* Firefox */
+      -ms-overflow-style: none; /* IE and Edge */
+      scrollbar-width: none; /* Firefox */
     }
   }
 }
